add read receipt func
